requirements.txt
Flask==3.0.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
piexif==1.1.3
Werkzeug==3.0.1 